+++ CHANGELOG.md
@@ -3,9 +3,10 @@ Changelog
 
 ## [0.5.0] — unreleased
 
--   Add `ConvApprox` and `CastApprox`
--   Support `Conv` and `ConvFloat` for arrays and tuples
--   Remove `impl<T> Conv<T> for T`
+-   Add `ConvApprox` and `CastApprox` (#21)
+-   Add `Result` type (#22)
+-   Support `Conv` and `ConvFloat` for arrays and tuples (#21)
+-   Remove `impl<T> Conv<T> for T` (#21)
 
 ## [0.4.4] — 2021-04-12